pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "meeting-notes-agent"
version = "0.1.0"
description = "Auto-detect Google Meet transcripts and produce summary, action items, Gmail draft, Asana tasks, and Notion entry — all draft, never sent."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Meridian Consulting Group" }]
keywords = ["claude", "anthropic", "google-meet", "meeting-notes", "automation"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
    "Topic :: Office/Business",
]

dependencies = [
    "anthropic>=0.50.0",
    "pydantic>=2.4.0",
    "python-dotenv>=1.0.0",
    "google-auth>=2.22.0",
    "google-auth-oauthlib>=1.1.0",
    "google-auth-httplib2>=0.1.1",
    "google-api-python-client>=2.100.0",
    "python-docx>=1.1.0",
    "notion-client>=2.2.1",
    "APScheduler>=3.10.4",
    "requests>=2.31.0",
    "python-dateutil>=2.8.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-mock>=3.11.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.4.0",
    "pyright>=1.1.350",
]

[project.scripts]
meeting-agent = "src.scheduler.main:main"

[project.urls]
Repository = "https://github.com/meridian/meeting-notes-agent"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

# ────────────────────────────────────────────────────────────────────────────
# Ruff — linter + formatter
# ────────────────────────────────────────────────────────────────────────────
[tool.ruff]
line-length = 100
target-version = "py310"
extend-exclude = [".venv", "build", "dist"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
    "SIM", # flake8-simplify
    "RUF", # ruff-specific
]
ignore = [
    "E501",   # line too long — formatter handles it
    "B008",   # function call in default arg
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["B017"]  # pytest.raises(Exception) is fine in tests

# ────────────────────────────────────────────────────────────────────────────
# Pyright — type checker
# ────────────────────────────────────────────────────────────────────────────
[tool.pyright]
include = ["src", "tests"]
exclude = ["**/__pycache__", ".venv"]
pythonVersion = "3.10"
typeCheckingMode = "basic"
reportMissingImports = "warning"
reportMissingTypeStubs = false

# ────────────────────────────────────────────────────────────────────────────
# pytest
# ────────────────────────────────────────────────────────────────────────────
[tool.pytest.ini_options]
minversion = "7.4"
testpaths = ["tests"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
]
filterwarnings = [
    "ignore::DeprecationWarning:google.*",
]
